third_party/blink/web_tests/external/wpt/app-history/navigate/navigate-transitionWhile-reject-event-order.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  let events = [];
  function finish() {
    assert_array_equals(events, ["onnavigate", "committed promisefulfilled", "onnavigateerror", "finished promiserejected"]);
    t.done();
  }

  appHistory.onnavigate = e => {
    events.push("onnavigate");
    e.transitionWhile(Promise.reject());
  };
  appHistory.onnavigatesuccess = () => events.push("onnavigatesuccess");
  appHistory.onnavigateerror = () => events.push("onnavigateerror");
  window.onhashchange = () => events.push("onhashchange");
  window.onpopstate = () => events.push("onpopstate");
  const result = appHistory.navigate("#1");

  result.committed.then(
      () => events.push("committed promisefulfilled"),
      () => events.push("committed promiserejected"));

  result.finished.then(
      () => events.push("finished promisefulfilled"),
      () => {
        events.push("finished promiserejected");
        t.step_timeout(finish, 0);
      });
}, "navigate() event ordering for same-document navigation");
</script>
